# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/agents": {
      "get": {
        "tags": [
          "agent-controller"
        ],
        "summary": "List agents",
        "operationId": "getApiAgents",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 50,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agent-controller"
        ],
        "summary": "Create agent",
        "operationId": "postApiAgents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{slug}/quotes": {
      "post": {
        "tags": [
          "quote-controller"
        ],
        "summary": "Create quote",
        "operationId": "postApiAgentsBySlugQuotes",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent-versions/{id}/dependencies": {
      "get": {
        "tags": [
          "dependency-controller"
        ],
        "summary": "List dependencies",
        "operationId": "getApiAgentVersionsByIdDependencies",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DependencyResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dependency-controller"
        ],
        "summary": "Create dependency",
        "operationId": "postApiAgentVersionsByIdDependencies",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDependencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent-versions/{id}/dependencies/{dependencyId}": {
      "delete": {
        "tags": [
          "dependency-controller"
        ],
        "summary": "Delete dependency",
        "operationId": "deleteApiAgentVersionsByIdDependenciesByDependencyId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "dependencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "dependency-controller"
        ],
        "summary": "Update dependency",
        "operationId": "patchApiAgentVersionsByIdDependenciesByDependencyId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "dependencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDependencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JsonNode": {},
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "traceId"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JsonNode"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "CreateAgentRequest": {
        "type": "object",
        "properties": {
          "developerId": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "semver": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "endpoint": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 0
          },
          "priceAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "network": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          },
          "asset": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          },
          "payTo": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          }
        },
        "required": [
          "asset",
          "description",
          "developerId",
          "endpoint",
          "name",
          "network",
          "payTo",
          "priceAtomic",
          "semver",
          "slug"
        ]
      },
      "AgentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "developerId": {
            "type": "string",
            "format": "uuid"
          },
          "developerName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentVersionResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "createdAt",
          "description",
          "developerId",
          "developerName",
          "id",
          "name",
          "slug",
          "updatedAt",
          "versions"
        ]
      },
      "AgentVersionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "semver": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "DISABLED"
            ]
          },
          "endpoint": {
            "type": "string",
            "format": "uri"
          },
          "priceAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "network": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "payTo": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "agentId",
          "asset",
          "createdAt",
          "endpoint",
          "id",
          "network",
          "payTo",
          "priceAtomic",
          "semver",
          "status",
          "updatedAt"
        ]
      },
      "QuoteRequest": {
        "type": "object",
        "properties": {
          "versionConstraint": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 128,
            "minLength": 0
          }
        }
      },
      "DependencySnapshot": {
        "type": "object",
        "properties": {
          "dependencyId": {
            "type": "string",
            "format": "uuid"
          },
          "targetAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "targetAgentSlug": {
            "type": "string"
          },
          "versionConstraint": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "maxPriceAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "maxCalls": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "resolved": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/QuoteSnapshot"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "dependencyId",
          "maxCalls",
          "maxPriceAtomic",
          "required",
          "targetAgentId",
          "targetAgentSlug",
          "versionConstraint"
        ]
      },
      "QuoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rootVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "maxCostAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "snapshot": {
            "$ref": "#/components/schemas/QuoteSnapshot"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteWarning"
            }
          }
        },
        "required": [
          "expiresAt",
          "id",
          "maxCostAtomic",
          "rootVersionId",
          "snapshot",
          "warnings"
        ]
      },
      "QuoteSnapshot": {
        "type": "object",
        "properties": {
          "version": {
            "$ref": "#/components/schemas/ResolvedVersionSnapshot"
          },
          "dependencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DependencySnapshot"
            }
          }
        },
        "required": [
          "dependencies",
          "version"
        ]
      },
      "QuoteWarning": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "dependencyId": {
            "type": "string",
            "format": "uuid"
          },
          "targetAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "targetAgentSlug": {
            "type": "string"
          },
          "versionConstraint": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "dependencyId",
          "targetAgentId",
          "targetAgentSlug",
          "versionConstraint"
        ]
      },
      "ResolvedVersionSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "agentSlug": {
            "type": "string"
          },
          "semver": {
            "type": "string"
          },
          "endpoint": {
            "type": "string",
            "format": "uri"
          },
          "priceAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "network": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "payTo": {
            "type": "string"
          }
        },
        "required": [
          "agentId",
          "agentSlug",
          "asset",
          "endpoint",
          "id",
          "network",
          "payTo",
          "priceAtomic",
          "semver"
        ]
      },
...
      "CreateDependencyRequest": {
        "type": "object",
        "properties": {
          "targetAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "versionConstraint": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          },
          "required": {
            "type": "boolean"
          },
          "maxPriceAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "maxCalls": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          }
        },
        "required": [
          "maxPriceAtomic",
          "targetAgentId",
          "versionConstraint"
        ]
      },
      "DependencyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "targetAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "targetAgentSlug": {
            "type": "string"
          },
          "versionConstraint": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "maxPriceAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "maxCalls": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "createdAt",
          "id",
          "maxCalls",
          "maxPriceAtomic",
          "required",
          "sourceVersionId",
          "targetAgentId",
          "targetAgentSlug",
          "updatedAt",
          "versionConstraint"
        ]
      },
...
      "UpdateDependencyRequest": {
        "type": "object",
        "properties": {
          "versionConstraint": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 128,
            "minLength": 0
          },
          "required": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "maxPriceAtomic": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[0-9]+$"
          },
          "maxCalls": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          }
        }
      },
...
      "AgentListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "items"
        ]
      }
    }
  }
}